custom queryset method to bulk update customer balance
